requirements.txt
# Telegram Bot
python-telegram-bot[webhooks]>=20.0

# Web Framework
Flask>=3.0.0
Flask-Login>=0.6.3
Werkzeug>=3.0.0

# Database
aiosqlite>=0.19.0

# HTTP Client
httpx>=0.25.0
aiohttp>=3.9.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0

# Security
bcrypt>=4.1.0

# Logging
structlog>=23.2.0
